+++ src/internal/connector/onedrive/item.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"io"
+	"strings"
 
 	msdrives "github.com/microsoftgraph/msgraph-sdk-go/drives"
 	"github.com/microsoftgraph/msgraph-sdk-go/models"
@@ -94,7 +95,7 @@ func driveItemReader(
 // doesn't have its size value updated as a side effect of creation,
 // and kiota drops any SetSize update.
 func oneDriveItemInfo(di models.DriveItemable, itemSize int64) *details.OneDriveInfo {
-	email := ""
+	var email, parent string
 
 	if di.GetCreatedBy().GetUser() != nil {
 		// User is sometimes not available when created via some
@@ -105,13 +106,21 @@ func oneDriveItemInfo(di models.DriveItemable, itemSize int64) *details.OneDrive
 		}
 	}
 
+	if di.GetParentReference() != nil {
+		if di.GetParentReference().GetName() != nil {
+			// EndPoint is not always populated from external apps
+			parent = *di.GetParentReference().GetName()
+		}
+	}
+
 	return &details.OneDriveInfo{
-		ItemType: details.OneDriveItem,
-		ItemName: *di.GetName(),
-		Created:  *di.GetCreatedDateTime(),
-		Modified: *di.GetLastModifiedDateTime(),
-		Size:     itemSize,
-		Owner:    email,
+		ItemType:  details.OneDriveItem,
+		ItemName:  *di.GetName(),
+		Created:   *di.GetCreatedDateTime(),
+		Modified:  *di.GetLastModifiedDateTime(),
+		DriveName: parent,
+		Size:      itemSize,
+		Owner:     email,
 	}
 }
 
@@ -120,10 +129,11 @@ func oneDriveItemInfo(di models.DriveItemable, itemSize int64) *details.OneDrive
 // separately for restore processes because the local itemable
 // doesn't have its size value updated as a side effect of creation,
 // and kiota drops any SetSize update.
+// TODO: Update drive name during Issue #2071
 func sharePointItemInfo(di models.DriveItemable, itemSize int64) *details.SharePointInfo {
 	var (
-		id  string
-		url string
+		id, parent, url string
+		reference       = di.GetParentReference()
 	)
 
 	// TODO: we rely on this info for details/restore lookups,
@@ -140,14 +150,29 @@ func sharePointItemInfo(di models.DriveItemable, itemSize int64) *details.ShareP
 		}
 	}
 
+	if reference != nil {
+		parent = *reference.GetDriveId()
+
+		if reference.GetName() != nil {
+			// EndPoint is not always populated from external apps
+			temp := *reference.GetName()
+			temp = strings.TrimSpace(temp)
+
+			if temp != "" {
+				parent = temp
+			}
+		}
+	}
+
 	return &details.SharePointInfo{
-		ItemType: details.OneDriveItem,
-		ItemName: *di.GetName(),
-		Created:  *di.GetCreatedDateTime(),
-		Modified: *di.GetLastModifiedDateTime(),
-		Size:     itemSize,
-		Owner:    id,
-		WebURL:   url,
+		ItemType:  details.OneDriveItem,
+		ItemName:  *di.GetName(),
+		Created:   *di.GetCreatedDateTime(),
+		Modified:  *di.GetLastModifiedDateTime(),
+		DriveName: parent,
+		Size:      itemSize,
+		Owner:     id,
+		WebURL:    url,
 	}
 }