+++ src/internal/operations/backup.go
@@ -207,8 +207,8 @@ func (op *BackupOperation) Run(ctx context.Context) (err error) {
 		return opStats.writeErr
 	}
 
-	// TODO: should always be 1, since backups are 1:1 with resourceOwners now.
-	opStats.resourceCount = len(data.ResourceOwnerSet(cs))
+	// should always be 1, since backups are 1:1 with resourceOwners.
+	opStats.resourceCount = 1
 	opStats.started = true
 	opStats.gc = gc.AwaitStatus()
 
@@ -218,6 +218,11 @@ func (op *BackupOperation) Run(ctx context.Context) (err error) {
 // checker to see if conditions are correct for incremental backup behavior such as
 // retrieving metadata like delta tokens and previous paths.
 func useIncrementalBackup(sel selectors.Selector, opts control.Options) bool {
+	// Delta-based incrementals currently only supported for Exchange
+	if sel.Service != selectors.ServiceExchange {
+		return false
+	}
+
 	return !opts.ToggleFeatures.DisableIncrementals
 }
 
@@ -344,10 +349,8 @@ func produceManifestsAndMetadata(
 			continue
 		}
 
-		tk, _ := kopia.MakeTagKV(kopia.TagBackupID)
-
-		bID := man.Tags[tk]
-		if len(bID) == 0 {
+		bID, ok := man.GetTag(kopia.TagBackupID)
+		if !ok {
 			return nil, nil, false, errors.New("snapshot manifest missing backup ID")
 		}
 
@@ -502,12 +505,9 @@ func consumeBackupDataCollections(
 	}
 
 	for _, reason := range reasons {
-		tags[reason.ResourceOwner] = ""
-
-		// TODO(ashmrtn): Create a separate helper function to go from service/cat
-		// to a tag.
-		serviceCat, _ := kopia.MakeServiceCat(reason.Service, reason.Category)
-		tags[serviceCat] = ""
+		for _, k := range reason.TagKeys() {
+			tags[k] = ""
+		}
 	}
 
 	bases := make([]kopia.IncrementalBase, 0, len(mans))
@@ -568,8 +568,10 @@ func mergeDetails(
 			continue
 		}
 
-		k, _ := kopia.MakeTagKV(kopia.TagBackupID)
-		bID := man.Tags[k]
+		bID, ok := man.GetTag(kopia.TagBackupID)
+		if !ok {
+			return errors.Errorf("no backup ID in snapshot manifest with ID %s", man.ID)
+		}
 
 		_, baseDeets, err := getBackupAndDetailsFromID(
 			ctx,